requirements.txt
flask>=3.0.0
python-vlc>=3.0.18121
pillow>=10.0.0
